refactor: product jpa로 리팩토링
