+++ src/main/java/gift/repository/ProductRepository.java
@@ -3,7 +3,7 @@ package gift.repository;
 import gift.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface ProductRepository extends JpaRepository<Product,Long> {
+public interface ProductRepository extends JpaRepository<Product,Long> {
 
 
 }
