+++ src/main/java/gift/repository/WishRepository.java
@@ -4,7 +4,7 @@ package gift.repository;
 import gift.entity.Wish;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface WishRepository extends JpaRepository<Wish,Long> {
+public interface WishRepository extends JpaRepository<Wish,Long> {
 
     public Wish findByMemberId(Long memberId);
